refactor: modularize non-fungible balances
